FIXED MY ISSUE YO!
The issue arose because the solvers/pulp appear to have an issue when setting values
greater than or equal to zero. Instead you need to introduce a very slight
perturbation to ensure that the duals will be reported appropriately.

Now I need to build the functionality around this.
Plots, DataFrames, all the fun stuff.
But it's working, and that ist gut. BOOM Fuckers